requirements.txt
Werkzeug[watchdog]==2.2.2 # https://github.com/pallets/werkzeug
ipdb==0.13.9  # https://github.com/gotcha/ipdb
psycopg2==2.9.3  # https://github.com/psycopg/psycopg2
watchfiles==0.17.0  # https://github.com/samuelcolvin/watchfiles

# Django
# ------------------------------------------------------------------------------
django==4.0.0
django-allauth==0.51.0
django-cors-headers==3.13.0
django-debug-toolbar==3.7.0
django-environ==0.9.0
django-extensions==3.2.1
djangorestframework==3.14.0
drf-spectacular==0.24.2
factory-boy==3.2.1
Faker==15.1.1
channels==3.0.4
channels_redis

# Testing
# ------------------------------------------------------------------------------
mypy==0.982  # https://github.com/python/mypy
django-stubs==1.12.0  # https://github.com/typeddjango/django-stubs
pytest==7.1.3  # https://github.com/pytest-dev/pytest
pytest-sugar==0.9.5  # https://github.com/Frozenball/pytest-sugar
djangorestframework-stubs==1.7.0  # https://github.com/typeddjango/djangorestframework-stubs
pytest-django
pytest-asyncio

# Code quality
# ------------------------------------------------------------------------------
flake8==5.0.4  # https://github.com/PyCQA/flake8
flake8-isort==4.2.0  # https://github.com/gforcada/flake8-isort
coverage==6.5.0  # https://github.com/nedbat/coveragepy
black==22.8.0  # https://github.com/psf/black
pylint-django==2.5.3  # https://github.com/PyCQA/pylint-django
pylint-celery==0.3  # https://github.com/PyCQA/pylint-celery
pre-commit==2.20.0  # https://github.com/pre-commit/pre-commit
